Fix for dupli type bit flags exceeding short range.
The nparticle dupli flag was using 1<<15, which gives issues with signed
short variables. Fortunately there are unused bits in the
Object.transflag, now reuse the deprecated OB_POWERTRACK flag for this.

Also replaced dupli context int type with short to avoid errors with
strict compiler flags.

+++ source/blender/blenkernel/intern/object_dupli.c
@@ -93,7 +93,7 @@ typedef struct DupliContext {
 } DupliContext;
 
 typedef struct DupliGenerator {
-	int type;				/* dupli type */
+	short type;				/* dupli type */
 	void (*make_duplis)(const DupliContext *ctx);
 } DupliGenerator;
